+++ packages/ui/src/components/chat/markdown/markdownTheme.ts
@@ -1,5 +1,4 @@
 import { registerCustomTheme, type ThemeRegistrationResolved } from '@pierre/diffs';
-import type { Theme } from '@/types/theme';
 import { MARKDOWN_SHIKI_THEME, MARKDOWN_SHIKI_THEME_DEFINITION } from './markdownShikiThemeDefinition';
 
 // The static Shiki theme name. Its definition (token colors referencing
@@ -26,30 +25,4 @@ export const ensureMarkdownShikiTheme = (): void => {
   registerCustomTheme(MARKDOWN_SHIKI_THEME, () =>
     Promise.resolve(MARKDOWN_SHIKI_THEME_DEFINITION as unknown as ThemeRegistrationResolved),
   );
-};
-
-/**
- * Build the `--md-syntax-*` CSS custom properties for the given app theme.
- * Apply the result as inline styles on the markdown container so the static
- * Shiki theme resolves to the active palette.
- */
-export const getMarkdownSyntaxVars = (theme: Theme): Record<string, string> => {
-  const base = theme.colors.syntax.base;
-  const tokens = theme.colors.syntax.tokens ?? {};
-  const status = theme.colors.status;
-
-  return {
-    '--md-syntax-foreground': base.foreground,
-    '--md-syntax-comment': base.comment,
-    '--md-syntax-string': base.string,
-    '--md-syntax-number': base.number,
-    '--md-syntax-keyword': base.keyword,
-    '--md-syntax-operator': base.operator,
-    '--md-syntax-function': base.function,
-    '--md-syntax-type': base.type,
-    '--md-syntax-variable': base.variable,
-    '--md-syntax-property': tokens.variableProperty ?? base.variable,
-    '--md-syntax-inserted': status.success,
-    '--md-syntax-deleted': status.error,
-  };
 };
